fix: improve components

+++ src/app/components/molecules/favouriteList/index.tsx
@@ -4,10 +4,11 @@ import style from "../largestCities/largest.module.scss";
 import FrostCard from "../../elements/card";
 import { useAppDispatch, useAppSelector } from "../../../store";
 import { sortCities } from "../../../helpers/sortCities";
-import { removeFromFavourites } from "../../../store/reducers/favourites/fav.reducer";
+import { removeFromFavourites } from "../../../store/reducers/favourites/favouriteSlice";
 import { useNavigate } from "react-router-dom";
 import { locationType } from "../../../store/reducers/types";
 import { getNoteKey } from "../../../helpers/getNoteKey";
+import { clearNotes } from "../../../store/reducers/notes/notesSlice";
 
 const FavouriteList = () => {
   const { favList } = useAppSelector((state) => state.favourites);
@@ -21,18 +22,21 @@ const FavouriteList = () => {
   };
 
   const removeFromList = (selected: locationType) => () => {
-    const noteKey = getNoteKey(selected);
-    if (notes[noteKey] && notes[noteKey].length) {
-      let proceedings = prompt(
-        `This location have notes attached to it\nType ${selected.location.name} to continue deletion.`,
-        ""
-      );
-      if (proceedings === selected.location.name) {
-        dispatch(removeFromFavourites({ data: selected }));
-      } else {
-        alert("Incomplete! name did not match.");
-      }
-    } else dispatch(removeFromFavourites({ data: selected }));
+    if (selected) {
+      const noteKey = getNoteKey(selected);
+      if (notes[noteKey] && notes[noteKey].length) {
+        let proceedings = prompt(
+          `This location have notes attached to it\nType ${selected.location.name} to continue deletion.`,
+          ""
+        );
+        if (proceedings === selected.location.name) {
+          dispatch(removeFromFavourites({ data: selected }));
+          dispatch(clearNotes({ key: noteKey }));
+        } else {
+          alert("Incomplete! name did not match.");
+        }
+      } else dispatch(removeFromFavourites({ data: selected }));
+    }
   };
 
   return (
@@ -43,12 +47,13 @@ const FavouriteList = () => {
       <div className={style.largest}>
         {sortCities(favList, "a-z").map((el, i) => (
           <FrostCard key={i} className={style.largest__card}>
-            <div
+            <button
               className={style.largest__card_remove}
+              data-testid={"remove-fav"}
               onClick={removeFromList(el)}
             >
               x
-            </div>
+            </button>
             <div onClick={gotoCity(el)}>
               <Typography level={"f32"}>{el?.current.temperature}°</Typography>
               <Typography level={"f12"} style={{ whiteSpace: "nowrap" }}>
